edge/pi/requirements.txt
requests==2.31.0
python-dotenv==1.0.1
numpy==1.26.4
pyserial==3.5
